pyproject.toml
[tool.poetry]
name = "voice-rag"
version = "0.1.0"
description = ""
authors = ["Andrea D'Agostino <andrea@dagtech.io>"]
readme = "README.md"
packages = [{include="src"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<=3.11"
upstash-vector = "^0.2.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.5"
langchain-community = "^0.0.17"
insanely-fast-whisper = "^0.0.13"
typer = "^0.9.0"
openai = "^1.11.0"
tiktoken = "^0.5.2"
ruff = "^0.2.0"
taipy = "^3.0.0"
rich = "^13.7.0"
pydub = "^0.25.1"
sqlmodel = "^0.0.14"
streamlit = "^1.31.0"
watchdog = "^4.0.0"
plotly = "^5.18.0"
graphviz = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
